feat(ui): move createPossession to a new file
